Add post on PHP-FPM configuration structure
This commit introduces a detailed blog post on the PHP-FPM config structure. From explaining the PHP-FPM config organization to using 'include' in configs, the post offers in-depth information on every query related to the PHP-FPM config structure. The post also includes appropriate visual and code examples for better comprehension. It becomes a useful guide for developers